Move countMatches to test project
